Add files via upload
##Crime Analysis Raipur city 

#About Project
Project designed to help the police to analyse the crime in the city.
This project is created for thesis purpose which shows the analysis of crime in the city of Raipur Chhattisgarh State.
In this project a real life data is collected and data analysis is performed using Python language. It shows how different features affect the crime rates and also shows the rates and numbers of crimes using graph and charts and location maps

Project started by looking at the data and ran some exploratory data analysis to understand the features and their relationships with eachother and with the target. 
At the next step, I preprocessed, cleaned and prepared the data for being used in fitting the models.
To choose a model, first considering that it was a classification problem I took a look on the classification models (Logistic_Regression, Decision_Tree, Random_Forest)


## Requirements

To run this code, you need to have :

* pandas
* numpy
* matplotlib
* seaborn
* Scikit-learn
* DBSCAN
* Folium


## Running instructions

Simply run the jupyter notebook


## Code structure
1. Import the data
2. Data visualization
3. Model training
4. Testing
5. Accuracy check

Dataset source: Chhattisgarh Police